requirements.txt
# Core dependencies
PyQt6>=6.6.1
openai>=1.66.0
tiktoken>=0.6.0

python-dotenv>=1.0.0

cachetools>=5.5.2
pytest>=8.3.5
Rx>=3.2.0
SQLAlchemy>=2.0.39
aiohttp>=3.11.11